basic components done and changed routes
